pyproject.toml
[project]
name = "kknaks-mobile"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "open-kknaks>=1.1.0",
    "python-dotenv>=1.2.2",
    "slack-bolt>=1.28.0",
]
